+++ python/ray/air/BUILD
@@ -201,22 +201,6 @@ py_test(
     deps = [":ml_lib"]
 )
 
-py_test(
-    name = "test_tensorflow_utils",
-    size = "small",
-    srcs = ["tests/test_tensorflow_utils.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"]
-)
-
-py_test(
-    name = "test_torch_utils",
-    size = "small",
-    srcs = ["tests/test_torch_utils.py"],
-    tags = ["team:ml", "exclusive"],
-    deps = [":ml_lib"]
-)
-
 # This is a dummy test dependency that causes the above tests to be
 # re-run if any of these files changes.
 py_library(
