+++ python/ray/train/sklearn/sklearn_predictor.py
@@ -4,9 +4,9 @@ import pandas as pd
 from joblib import parallel_backend
 from sklearn.base import BaseEstimator
 
-from ray.air._internal.sklearn_utils import set_cpu_params
 from ray.air.checkpoint import Checkpoint
 from ray.air.constants import TENSOR_COLUMN_NAME
+from ray.train._internal.sklearn_utils import set_cpu_params
 from ray.train.predictor import Predictor
 from ray.train.sklearn import load_checkpoint
 from ray.util.joblib import register_ray
