+++ python/ray/train/sklearn/sklearn_trainer.py
@@ -4,32 +4,31 @@ import warnings
 from collections import defaultdict
 from time import time
 from traceback import format_exc
-from typing import Any, Callable, Dict, Iterable, Optional, Tuple, Union, TYPE_CHECKING
+from typing import TYPE_CHECKING, Any, Callable, Dict, Iterable, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
 from joblib import parallel_backend
-
-from ray import tune
-import ray.cloudpickle as cpickle
-from ray.air.checkpoint import Checkpoint
-from ray.air.config import RunConfig, ScalingConfig
-from ray.train.constants import MODEL_KEY, TRAIN_DATASET_KEY
-from ray.train.trainer import GenDataset, BaseTrainer
-from ray.air._internal.checkpointing import (
-    load_preprocessor_from_dir,
-    save_preprocessor_to_dir,
-)
-from ray.air._internal.sklearn_utils import has_cpu_params, set_cpu_params
-from ray.util import PublicAPI
-from ray.util.joblib import register_ray
-
 from sklearn.base import BaseEstimator, clone
 from sklearn.metrics import check_scoring
 from sklearn.model_selection import BaseCrossValidator, cross_validate
 
 # we are using a private API here, but it's consistent across versions
 from sklearn.model_selection._validation import _check_multimetric_scoring, _score
+
+import ray.cloudpickle as cpickle
+from ray import tune
+from ray.air._internal.checkpointing import (
+    load_preprocessor_from_dir,
+    save_preprocessor_to_dir,
+)
+from ray.air.checkpoint import Checkpoint
+from ray.air.config import RunConfig, ScalingConfig
+from ray.train._internal.sklearn_utils import has_cpu_params, set_cpu_params
+from ray.train.constants import MODEL_KEY, TRAIN_DATASET_KEY
+from ray.train.trainer import BaseTrainer, GenDataset
+from ray.util import PublicAPI
+from ray.util.joblib import register_ray
 
 if TYPE_CHECKING:
     from ray.data.preprocessor import Preprocessor
