+++ python/ray/train/tests/test_tensorflow_utils.py
@@ -3,7 +3,7 @@ import pandas as pd
 import tensorflow as tf
 
 from ray.data.extensions import TensorArray
-from ray.air._internal.tensorflow_utils import convert_pandas_to_tf_tensor
+from ray.train._internal.tensorflow_utils import convert_pandas_to_tf_tensor
 
 
 def test_convert_simple_df_to_tensor():
