+++ python/ray/train/torch/torch_predictor.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import torch
 
-from ray.air._internal.torch_utils import convert_pandas_to_torch_tensor
 from ray.air.checkpoint import Checkpoint
+from ray.train._internal.torch_utils import convert_pandas_to_torch_tensor
 from ray.train.predictor import DataBatchType, Predictor
 from ray.train.torch import load_checkpoint
