+++ python/ray/train/torch/torch_trainer.py
@@ -1,12 +1,13 @@
-from typing import Callable, Optional, Dict, Tuple, Union, TYPE_CHECKING
+from typing import TYPE_CHECKING, Callable, Dict, Optional, Tuple, Union
+
 import torch
 
+from ray.air.checkpoint import Checkpoint
+from ray.air.config import DatasetConfig, RunConfig, ScalingConfig
+from ray.train._internal.torch_utils import load_torch_model
+from ray.train.data_parallel_trainer import DataParallelTrainer, _load_checkpoint
 from ray.train.torch.config import TorchConfig
 from ray.train.trainer import GenDataset
-from ray.train.data_parallel_trainer import DataParallelTrainer, _load_checkpoint
-from ray.air.config import ScalingConfig, RunConfig, DatasetConfig
-from ray.air.checkpoint import Checkpoint
-from ray.air._internal.torch_utils import load_torch_model
 from ray.util import PublicAPI
 
 if TYPE_CHECKING:
